Separate activity data from retrieve activity servlet.
